1.增加 可分享APK文件到安装狮来发起安装(考虑到大部分定制系统的文件管理无法直接使用第三方应用发起安装) 2.安装狮-自定义 增加处理卸载请求

+++ app/src/main/java/com/modosa/apkinstaller/activity/Install6Activity.java
@@ -2,6 +2,7 @@ package com.modosa.apkinstaller.activity;
 
 import android.content.ComponentName;
 import android.content.Intent;
+import android.net.Uri;
 import android.util.Log;
 
 import androidx.preference.PreferenceManager;
@@ -18,20 +19,44 @@ import java.io.File;
  */
 public class Install6Activity extends AbstractInstallerActivity {
 
+    private Intent startIntent;
+    private boolean needDelete = false;
+
+
     @Override
     public void startInstall(String getinstallApkPath) {
-        installApkFile = new File(getinstallApkPath);
-        boolean needDelete = istemp;
-        istemp = false;
-        Log.d("Start install", getinstallApkPath + "");
+        startOperate(true, getinstallApkPath);
+    }
 
-        boolean prepareInstall = false;
+    @Override
+    protected void startUninstall(String uninstallPkgname) {
+        startOperate(false, uninstallPkgname);
+    }
+
+
+    private void startOperate(boolean isInstall, String param) {
+
+        if (isInstall) {
+            installApkFile = new File(param);
+            needDelete = istemp;
+            istemp = false;
+            Log.d("Start install", param + "");
+        } else {
+            Log.d("Start uninstall", param + "");
+        }
+
+        boolean prepareOperate = false;
         String getCustomInstaller = PreferenceManager.getDefaultSharedPreferences(this).getString(MainFragment.SP_KEY_CUSTOM_INSTALLER, "").trim();
         if (!"".equals(getCustomInstaller)) {
-            Intent intent = new Intent(Intent.ACTION_VIEW)
-                    .setDataAndType(OpUtil.getMyContentUriForFile(this, installApkFile), "application/vnd.android.package-archive")
-                    .setFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                    .addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            if (isInstall) {
+                startIntent = new Intent(Intent.ACTION_VIEW)
+                        .setDataAndType(OpUtil.getMyContentUriForFile(this, installApkFile), "application/vnd.android.package-archive")
+                        .setFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+                        .addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+            } else {
+                startIntent = new Intent(Intent.ACTION_DELETE, Uri.parse("package:" + param))
+                        .setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            }
 
             if (getCustomInstaller.contains("/")) {
                 try {
@@ -39,20 +64,25 @@ public class Install6Activity extends AbstractInstallerActivity {
                     if (names[1].length() > 1 && names[1].startsWith(".")) {
                         names[1] = names[0] + names[1];
                     }
-                    intent.setComponent(new ComponentName(names[0], names[1]));
-                    prepareInstall = true;
+                    startIntent.setComponent(new ComponentName(names[0], names[1]));
+                    prepareOperate = true;
                 } catch (Exception ignore) {
                 }
             } else {
-                intent.setPackage(getCustomInstaller);
-                prepareInstall = true;
+                startIntent.setPackage(getCustomInstaller);
+                prepareOperate = true;
             }
 
-            if (prepareInstall) {
+            if (prepareOperate) {
                 try {
                     String installerLable = AppInfoUtil.getCustomInstallerLable(this, getCustomInstaller);
-                    showMyToast0(String.format(getString(R.string.tip_use_to_install), installerLable, apkinfo[0]));
-                    startActivity(intent);
+                    if (isInstall) {
+                        showMyToast0(String.format(getString(R.string.tip_use_to_install), installerLable, apkinfo[0]));
+                    } else {
+                        showMyToast0(String.format(getString(R.string.tip_use_to_uninstall), installerLable, uninstallPackageLable));
+                    }
+                    Log.e("intent0", ": " + startIntent);
+                    startActivity(startIntent);
                     needDelete = false;
                 } catch (Exception e) {
                     showMyToast1(String.format(getString(R.string.tip_invalid__custom_installer), getCustomInstaller));
@@ -64,13 +94,11 @@ public class Install6Activity extends AbstractInstallerActivity {
         } else {
             showMyToast1(R.string.tip_not_set_custom_installer);
         }
-        if (needDelete) {
+
+        if (isInstall && needDelete) {
             OpUtil.deleteSingleFile(installApkFile);
         }
         finish();
     }
 
-    @Override
-    protected void startUninstall(String uninstallPkgname) {
-    }
 }
